finished contact pages. added roles and SEO tags

+++ routes/web.php
@@ -34,12 +34,12 @@ Route::get('/admin', 'admin\AdminController@dashboard');
 
 
 //Admin Food Categories
-Route::get('/admin/food-categories', 'admin\FoodCategoriesController@index');
-Route::get('/admin/food-categories/create', 'admin\FoodCategoriesController@create');
-Route::post('/admin/food-categories', 'admin\FoodCategoriesController@store');
-Route::get('/admin/food-categories/{id}/edit', 'admin\FoodCategoriesController@edit');
-Route::put('/admin/food-categories/{id}', 'admin\FoodCategoriesController@update');
-Route::delete('/admin/food-categories/{id}/delete', 'admin\FoodCategoriesController@delete');
+Route::get('/admin/food-categories', 'admin\FoodCategoriesController@index')->middleware('role:Admin');
+Route::get('/admin/food-categories/create', 'admin\FoodCategoriesController@create')->middleware('role:Admin');
+Route::post('/admin/food-categories', 'admin\FoodCategoriesController@store')->middleware('role:Admin');
+Route::get('/admin/food-categories/{id}/edit', 'admin\FoodCategoriesController@edit')->middleware('role:Admin');
+Route::put('/admin/food-categories/{id}', 'admin\FoodCategoriesController@update')->middleware('role:Admin');
+Route::delete('/admin/food-categories/{id}/delete', 'admin\FoodCategoriesController@delete')->middleware('role:Admin');
 
 
 //Admin Food Items
@@ -61,12 +61,12 @@ Route::post('/admin/reservations', 'admin\ReservationsController@store');
 Route::delete('/admin/reservations/{id}/delete', 'admin\ReservationsController@delete');
 
 //Admin Users
-Route::get('/admin/users', 'admin\UsersController@index');
-Route::get('/admin/users/create', 'admin\UsersController@create');
-Route::post('/admin/users', 'admin\UsersController@store');
-Route::get('/admin/users/{id}/edit', 'admin\UsersController@edit');
-Route::put('/admin/users/{id}', 'admin\UsersController@update');
-Route::delete('/admin/users/{id}/delete', 'admin\UsersController@delete');
+Route::get('/admin/users', 'admin\UsersController@index')->middleware('role:Admin');
+Route::get('/admin/users/create', 'admin\UsersController@create')->middleware('role:Admin');
+Route::post('/admin/users', 'admin\UsersController@store')->middleware('role:Admin');
+Route::get('/admin/users/{id}/edit', 'admin\UsersController@edit')->middleware('role:Admin');
+Route::put('/admin/users/{id}', 'admin\UsersController@update')->middleware('role:Admin');
+Route::delete('/admin/users/{id}/delete', 'admin\UsersController@delete')->middleware('role:Admin');
 
 //Admin Settings
 Route::get('/admin/settings/general', 'admin\SettingController@general');
